Clarifying existence of `opts` option in createNetworkInterface
I had a little bit of trouble migrating from a previous version of Apollo Client where the `fetch` options weren't an explicit key in the `createNetworkInterface` function. I added some clarifying language in the `createBatchingNetworkInterface` part to make sure it was clear that one additional option was available (along with the other options from `createNetworkInterface`) and left an example in both parts to make it extremely clear.